requirements.txt
pywin32==224
requests==2.21.0
pillow==5.3.0
selenium==3.141.0
baidu-aip==2.2.10.0
